Created cron dir; Extracted some lib funtionality; Refactored directory structure
