+++ lib/machine.py
