+++ tools/README.md
@@ -1,0 +1,9 @@
+## What is in here?
+
+The tools folder contains all scripts that are mainly used as stand-alone CLI programs.
+
+For instance `prune_db.py` to clear the DB manually.
+
+Some files have also counterparts in the *lib* directory where they include functionality from.
+An example would be the `phase_stats.py` which is often included, but in some cases also needs a *tools* CLI to be
+invoked manually.
